Overhauled Analyzer. Analysis is better-factored into serveral policies -- spectral peak selection, partial formation, and bandwidth association -- making experimentation much easier.
The Partial formation policy has been modified to follow an optional
frequency reference envelope, which is used t warp the spectrum (just
the selected peaks actually) before matching peaks and forming partials.
At the end of the analysis, the Partial frequencies are unwarped. This
allows the Analyzer to do a much better job of tracking a known pitch
contour when constructing Partials. An optional argument has been added
to Analyzer::analyze(), corresponding to the reference envelope.

This tracking analysis is not yet available in the scripting interface
or the procedural interface.